requirements.txt
chromadb
PyPDF2
sentence-transformers
torch
tqdm
scikit-learn
openai
python-dotenv
transformers